refactor(embedding): split inference pipeline
